# contract/openapi.spec.yaml
openapi: 3.0.1
info:
  description: "API for submitting questionnaires with an array of answers"
  title: Questionnaire API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /questionnaires:
    get:
      summary: Get all questionnaires
      description: Retrieve all questionnaires
      responses:
        "200":
          description: A list of questionnaires
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Questionnaire"
        "500":
          description: Internal server error
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireSubmission"
        required: true
      responses:
        "204":
          description: Questionnaire submitted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid input data
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error
      summary: Submit Questionnaire
  /questions:
    get:
      summary: Get all questions
      description: Retrieve all questions, optionally filtered by questionnaire_id
      parameters:
        - name: questionnaire_id
          in: query
          description: ID of the questionnaire to filter questions
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A list of questions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Question"

        "400":
          description: Invalid parameter
        "500":
          description: Internal server error
components:
  schemas:
    Answer:
      type: object
      properties:
        id:
          type: string
          format: objectid
          title: ID
          readOnly: true
        question_id:
          type: string
          format: objectid
          title: Question ID
        text:
          title: Answer
        created_at:
          type: string
          format: date-time
          title: Created At
          readOnly: true
      required:
        - question_id
        - text
    AnswerDB:
      allOf:
        - $ref: "#/components/schemas/Answer"
        - type: object
          properties:
            user_id:
              type: string
              title: User ID
      required:
        - user_id
    Question:
      type: object
      properties:
        id:
          type: string
          format: objectid
          title: ID
          readOnly: true
        questionnaire_id:
          type: string
          format: objectid
          title: Questionnaire ID
        text:
          title: Text
        created_at:
          type: string
          format: date-time
          title: Created At
      required:
        - questionnaire_id
        - text
    Questionnaire:
      type: object
      properties:
        id:
          type: string
          format: objectid
          title: ID
          readOnly: true
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
          readOnly: true
      required:
        - name
    QuestionnaireSubmission:
      type: object
      properties:
        questionnaire_id:
          type: string
          format: objectid
          title: Questionnaire ID
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
      required:
        - questionnaire_id
        - answers
    APIResponse:
      type: object
      properties:
        data:
          title: Data
      required:
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
          title: Status
        message:
          type: string
          title: Message
      required:
        - status
        - message