# src/openapi_server/models/answer.py
# coding: utf-8

"""
    Questionnaire API

    API for submitting questionnaires with an array of answers

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json


from datetime import datetime
from pydantic import BaseModel, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional

from src.openapi_server.models.pyobjectid import PyObjectId

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class Answer(BaseModel):
    """
    Answer
    """  # noqa: E501

    id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
    question_id: PyObjectId
    text: Optional[Any]
    created_at: Optional[datetime] = Field(default_factory=datetime.utcnow)
    __properties: ClassVar[List[str]] = ["_id", "question_id", "answer", "created_at"]

    model_config = {
        "populate_by_name": True,
        "arbitrary_types_allowed": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> "Answer":
        """Create an instance of Answer from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        * OpenAPI `readOnly` fields are excluded.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
                "_id",
                "created_at",
            },
            exclude_none=True,
        )
        if self.answer is None and "answer" in self.model_fields_set:
            _dict["answer"] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> "Answer":
        """Create an instance of Answer from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate(
            {
                "_id": obj.get("_id"),
                "question_id": obj.get("question_id"),
                "answer": obj.get("answer"),
                "created_at": obj.get("created_at"),
            }
        )
        return _obj